requirements.txt
aiohttp
aiosignal
async-timeout
attrs
autopep8
blobfile
certifi
charset-normalizer
filelock
frozenlist
idna
lxml
multidict
numpy
openai
pycodestyle
pycryptodomex
python-dotenv
regex
requests
tiktoken
tqdm
urllib3
yarl
rich
ollama
yaspin